pyproject.toml
[project]
name = "nlp-insights"
version = "0.1.0"
description = "Clinical trials analysis and insights extraction"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "streamlit>=1.28.0",
    "nltk>=3.8.1",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "ipykernel>=6.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
